Refactor `LogIn` component to avoid circular deps
