+++ src/test/pages/LogIn.cy.tsx
@@ -3,7 +3,7 @@ import { LogIn } from "../../app/components/pages/LogIn";
 
 it('Log In page should have no visual regressions', () => {
     // Arrange
-    cy.mountWithStoreAndRouter(<LogIn/>, ["/"]);
+    cy.mountWithStoreAndRouter(<LogIn />, ["/login"]);
         
     // Assert
     cy.matchImage();
